Add files via upload
Chapter 6. PC #5 and #6 Combined and modified. Homework Assignment (read instructions carefully)

Write a program that asks the user for the name of a file. Assume that the file contains a series of decimal numbers, one on each line. The program should read all the numbers from the given file, calculate their total and average, and then display the total and average of all numbers in the following format (three decimal digits):


Total: nnnnn.nnn
Average: nnnnn.nnn